# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Branch {
  id: number;
  name: string;
  code: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface BranchInput {
  name: string;
  code: string;
  description?: string;
}

export interface BranchUpdate {
  name?: string;
  code?: string;
  description?: string;
}

export interface Subject {
  id: number;
  name: string;
  code: string;
  branchId: number;
  /** @nullable */
  branchName?: string | null;
  createdAt: string;
}

export interface SubjectInput {
  name: string;
  code: string;
  branchId: number;
}

export interface SubjectUpdate {
  name?: string;
  code?: string;
  branchId?: number;
}

export interface Student {
  id: number;
  studentId: string;
  name: string;
  branchId: number;
  /** @nullable */
  branchName?: string | null;
  year: number;
  section: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface StudentInput {
  studentId: string;
  name: string;
  branchId: number;
  year: number;
  section: string;
  phone?: string;
  email?: string;
}

export interface StudentUpdate {
  studentId?: string;
  name?: string;
  branchId?: number;
  year?: number;
  section?: string;
  phone?: string;
  email?: string;
}

export interface AttendanceRecord {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  studentCode?: string | null;
  subjectId: number;
  /** @nullable */
  subjectName?: string | null;
  branchId: number;
  /** @nullable */
  branchName?: string | null;
  date: string;
  /** @nullable */
  time?: string | null;
  status: string;
  createdAt: string;
}

export interface AttendanceRecordInput {
  studentId: number;
  subjectId: number;
  branchId: number;
  date: string;
  time?: string;
  status: string;
}

export interface AttendanceBulkInput {
  records: AttendanceRecordInput[];
}

export interface AttendanceUpdate {
  status?: string;
  time?: string;
}

export interface AttendanceListResponse {
  records: AttendanceRecord[];
  total: number;
  page: number;
  limit: number;
}

export interface AttendanceReport {
  collegeName: string;
  branchName: string;
  subjectName: string;
  subjectCode?: string;
  date: string;
  /** @nullable */
  facultyName?: string | null;
  totalStudents: number;
  totalPresent: number;
  totalAbsent: number;
  attendancePercentage: number;
  records: AttendanceRecord[];
}

export interface DashboardStats {
  totalStudents: number;
  totalBranches: number;
  totalSubjects: number;
  presentToday: number;
  absentToday: number;
  attendancePercentageToday: number;
  totalAttendanceRecords: number;
  overallAttendancePercentage: number;
}

export interface BranchStat {
  branchId: number;
  branchName: string;
  branchCode?: string;
  studentCount: number;
  presentCount: number;
  absentCount: number;
  attendancePercentage: number;
}

export interface AttendanceTrendPoint {
  date: string;
  present: number;
  absent: number;
  total: number;
}

export interface SubjectStat {
  subjectId: number;
  subjectName: string;
  subjectCode: string;
  branchId: number;
  branchName: string;
  presentCount: number;
  absentCount: number;
  attendancePercentage: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListSubjectsParams = {
  branchId?: number;
};

export type ListStudentsParams = {
  branchId?: number;
  year?: number;
  section?: string;
  search?: string;
};

export type ListAttendanceParams = {
  branchId?: number;
  subjectId?: number;
  studentId?: number;
  date?: string;
  startDate?: string;
  endDate?: string;
  status?: string;
  search?: string;
  sortOrder?: string;
  page?: number;
  limit?: number;
};

export type GetAttendanceReportParams = {
  branchId: number;
  subjectId: number;
  date?: string;
  startDate?: string;
  endDate?: string;
  facultyName?: string;
};

export type GetSubjectStatsParams = {
  branchId?: number;
};