# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BTI Smart Attendance Monitoring System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "404":
          description: Not found
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: studentId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: sortOrder
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceListResponse"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create attendance record(s)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBulkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/report:
    get:
      operationId: getAttendanceReport
      tags: [attendance]
      summary: Get attendance report for print/export
      parameters:
        - name: branchId
          in: query
          required: true
          schema:
            type: integer
        - name: subjectId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: facultyName
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Attendance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceReport"
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete an attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    Branch:
      type: object
      required: [id, name, code, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name, code]
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
...
    AttendanceRecord:
      type: object
      required: [id, studentId, subjectId, branchId, date, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        studentCode:
          type: ["string", "null"]
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    AttendanceRecordInput:
      type: object
      required: [studentId, subjectId, branchId, date, status]
      properties:
        studentId:
          type: integer
        subjectId:
          type: integer
        branchId:
          type: integer
        date:
          type: string
        time:
          type: string
        status:
          type: string
    AttendanceBulkInput:
      type: object
      required: [records]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecordInput"
    AttendanceUpdate:
      type: object
      properties:
        status:
          type: string
        time:
          type: string
    AttendanceListResponse:
      type: object
      required: [records, total, page, limit]
      properties:
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AttendanceReport:
      type: object
      required: [collegeName, branchName, subjectName, date, totalStudents, totalPresent, totalAbsent, attendancePercentage, records]
      properties:
        collegeName:
          type: string
        branchName:
          type: string
        subjectName:
          type: string
        subjectCode:
          type: string
        date:
          type: string
        facultyName:
          type: ["string", "null"]
        totalStudents:
          type: integer
        totalPresent:
          type: integer
        totalAbsent:
          type: integer
        attendancePercentage:
          type: number
        records:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceRecord"
...