# lib/api-zod/src/generated/types/attendanceListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecord } from "./attendanceRecord";

export interface AttendanceListResponse {
  records: AttendanceRecord[];
  total: number;
  page: number;
  limit: number;
}