# lib/api-zod/src/generated/types/attendanceReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecord } from "./attendanceRecord";

export interface AttendanceReport {
  collegeName: string;
  branchName: string;
  subjectName: string;
  subjectCode?: string;
  date: string;
  /** @nullable */
  facultyName?: string | null;
  totalStudents: number;
  totalPresent: number;
  totalAbsent: number;
  attendancePercentage: number;
  records: AttendanceRecord[];
}