# lib/api-zod/src/generated/types/branchUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BTI Smart Attendance Monitoring System API
 * OpenAPI spec version: 0.1.0
 */

export interface BranchUpdate {
  name?: string;
  code?: string;
  description?: string;
}